% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compare_simreal}
\alias{compare_simreal}
\title{#' @import ggplot2
#' @export}
\usage{
compare_simreal(real, simlist, plotdir, realcellinfo = NULL,
  realdegeneinfo = NULL)
}
\description{
#' @import ggplot2
#' @export
}
