% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{escoEstBCV}
\alias{escoEstBCV}
\title{Estimate esco Biological Coefficient of Variation parameters}
\usage{
escoEstBCV(counts, norm.counts, params, cellinfo = NULL)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{norm.counts}{normalized counts matrix to estimate parameters from.}

\item{params}{escoParams object to store estimated values in.}

\item{cellinfo}{info about the identity of each cell in the cell structure.
If cellinfo is not null, then BCV is 
estiamted adjusted to the cell structures}
}
\value{
escoParams object with estimated values.
}
\description{
Parameters are estimated using the \code{\link[edgeR]{estimateDisp}} function
in the \code{edgeR} package.
}
\details{
The \code{\link[edgeR]{estimateDisp}} function is used to estimate the common
dispersion and prior degrees of freedom. See
\code{\link[edgeR]{estimateDisp}} for details. When estimating parameters on
simulated data we found a broadly linear relationship between the true
underlying common dispersion and the \code{edgR} estimate, therefore we
apply a small correction, \code{disp = 0.1 + 0.25 * edgeR.disp}.
}
