% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{escoEstCorr}
\alias{escoEstCorr}
\title{Estimate esco correlation matrix using meta data or the copula method}
\usage{
escoEstCorr(counts, cellinfo, params, method = c("meta", "copula"))
}
\description{
Estimate esco correlation matrix using meta data or the copula method
}
