% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{escoEstDE}
\alias{escoEstDE}
\title{Estimate esco expression DE parameters for discrete cell groups}
\usage{
escoEstDE(nonDE.normcounts, DE.normcounts, degenes, cellinfo, params)
}
\arguments{
\item{nonDE.normcounts}{library size normalised counts matrix for nonDE genes.}

\item{DE.normcounts}{library size normalised counts matrix for DE genes.}

\item{degenes}{a dataframe of two columns: "genes" and "clust", 
where "genes" contains the position of a degenes, and "clust" contains the 
corresponding cell group the degenes marks.}

\item{cellinfo}{a vector of length n, where n is the number of cells. 
Each entries is the group identity of a cell.}

\item{params}{\code{escoParams} object to store estimated values in.}
}
\value{
\code{escoParams} object with estimated values.
}
\description{
Parameters are estimated by comparing means of individual genes to the
median mean expression level.
}
\details{
Expression DE factor are detected using the Median Absolute Deviation
(MAD) from median method. If the log2 mean expression of a gene is greater
than two MADs above the median log2 mean expression it is designated as an
outlier. The proportion of outlier genes is used to estimate the outlier
probability. Factors for each outlier gene are calculated by dividing mean
expression by the median mean expression. A log-normal distribution is then
fitted to these factors in order to estimate the outlier factor location and
scale parameters using \code{\link[fitdistrplus]{fitdist}}.
}
