% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{escoEstLib}
\alias{escoEstLib}
\title{Estimate esco library size parameters}
\usage{
escoEstLib(counts, norm.counts, params)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{norm.counts}{library size normalised counts matrix.}

\item{params}{escoParams object to store estimated values in.}
}
\value{
escoParams object with estimated values.
}
\description{
The Shapiro-Wilks test is used to determine if the library sizes are
normally distributed. If so a normal distribution is fitted to the library
sizes, if not (most cases) a log-normal distribution is fitted and the
estimated parameters are added to the params object. See
\code{\link[fitdistrplus]{fitdist}} for details on the fitting.
}
