% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{escoEstMean}
\alias{escoEstMean}
\title{Estimate esco group mean parameters}
\usage{
escoEstMean(normcounts, counts, params)
}
\arguments{
\item{params}{escoParams object to store estimated values in.}

\item{norm.counts}{library size normalised counts matrix.}
}
\value{
escoParams object with estimated values.
}
\description{
Estimate rate and shape parameters for the gamma distribution used to
simulate gene expression means.
}
\details{
Parameter for the gamma distribution are estimated by fitting the mean
normalised counts using \code{\link[fitdistrplus]{fitdist}}. The 'maximum
goodness-of-fit estimation' method is used to minimise the Cramer-von Mises
distance. This can fail in some situations, in which case the 'method of
moments estimation' method is used instead. Prior to fitting the means are
winsorized by setting the top and bottom 10 percent of values to the 10th
and 90th percentiles.
}
