% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn.R
\name{gcn}
\alias{gcn}
\title{Compute the gene co-expression network}
\usage{
gcn(count, genes = NULL, CPM = TRUE, CPM2 = FALSE,
  name = "pearson")
}
\arguments{
\item{count}{a gene by cell matrix containing the expression count.}

\item{genes}{a vector of names of selected genes, 
default is NULL, indicating choosing all the genes.}

\item{CPM}{whether to use Counts Per Millon 
normalization on the whole data sets or not.}

\item{CPM2}{whether to use Counts Per Millon 
normalization on the selected genes or not.}

\item{name}{a string indicating what type of 
correlation metric ("pearson", "spearman", "kendall","cosine") to use, 
default options is "pearson".}
}
\value{
a correlation matrix
}
\description{
compute the gene co-expression netowrk using different correlation metric
}
\examples{
data = matrix(rnorm(100),20,5)
gcndata = gcn(data)
}
