% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimDropout}
\alias{escoSimDropout}
\title{Simulate dropout}
\usage{
escoSimDropout(sim, trial, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add dropout to.}
}
\value{
SingleCellExperiment with simulated dropout and observed counts.
}
\description{
A logistic function is used to form a relationshop between the expression
level of a gene and the probability of dropout, giving a probability for each
gene in each cell. These probabilities are used in a Bernoulli distribution
to decide which counts should be dropped.
}
