% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimGroupCellMeans}
\alias{escoSimGroupCellMeans}
\title{Simulate group cell means}
\usage{
escoSimGroupCellMeans(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add cell means to.}

\item{verbose}{whether to print the process or not.}
}
\value{
SingleCellExperiment with added cell means.
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell where cells are of discrete group structure. 
For a cell we start with the base gene mean and then combined with the DE factor 
for the group/branch they belong to
(when simulating group/tree) if the gene is DE gene, finally the
means are adjusted for each cell's expected library size.
}
