% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimGroupDE}
\alias{escoSimGroupDE}
\title{Simulate differential expression (DE) factors for discrete groups}
\usage{
escoSimGroupDE(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add differential expression to.}

\item{verbose}{logical. Whether to print progress messages.}
}
\value{
SingleCellExperiment with simulated differential expression.
}
\description{
DE factors for each group are produced using \code{\link{getLNormFactors}} and these are added
along with updated means for each group.
}
