% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimLibSizes}
\alias{escoSimLibSizes}
\title{Simulate library sizes}
\usage{
escoSimLibSizes(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add library size to.}
}
\value{
SingleCellExperiment with simulated library sizes.
}
\description{
Simulate expected library sizes. Typically a log-normal distribution is used
but there is also the option to use a normal distribution. In this case any
negative values are set to half the minimum non-zero value.
}
