% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimSingleCellMeans}
\alias{escoSimSingleCellMeans}
\title{Simulate cell means}
\usage{
escoSimSingleCellMeans(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add cell means to.}

\item{verbose}{whether to print the process or not.}
}
\value{
SingleCellExperiment with added cell means.
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell where cells are of one single homonogenrous group. 
Cells start with the base gene mean and then adjusted for each cell's expected library size.
}
