% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimTrajDE}
\alias{escoSimTrajDE}
\title{Simulate differential expression (DE) factors for trajectory}
\usage{
escoSimTrajDE(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add differential expression to.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SingleCellExperiment with simulated trajetory means.
}
\description{
The DE factor for DE genes are simulated for each step along each path (branches of trajectory)
}
\details{
The method of simulating paths is inspired by the method used in the PROSSTT
simulation. Changes in expression are controlled by \code{paths.changes}
regulatory programs. For each path a random walk through \code{paths.diffusion} 
and \code{paths.velocity} is generated for each DE genes. 

The path structure itself is specified by the \code{paths.design} parameter.
This is a \code{data.frame} with three columns: "Path", "From", and "Steps".
The Path field is an ID for each path while the Steps field controls the
length of each path. Increasing the number of steps will increase the
difference in expression between the ends of the paths. The From field sets
the originating point of each path. For example a From of \code{0, 0, 0}
would indicate three paths from the origin while a From of \code{0, 1, 1}
would give a branching structure with Path 1 beginning at the origin and
Path 2 and Path 3 beginning at the end of Path 1.
}
\references{
Papadopoulos N, Parra RG, Söding J. PROSSTT: probabilistic simulation of
single-cell RNA-seq data for complex differentiation processes.
Bioinformatics (2019). \url{https://doi.org/10.1093/bioinformatics/btz078}.
}
