% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimTrajCellMeans}
\alias{escoSimTrajCellMeans}
\title{Simulate esco trajetory cell means}
\usage{
escoSimTrajCellMeans(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment containing simulation.}

\item{verbose}{logical. Whether to print progress messages.}
}
\value{
sim SingleCellExperiment with cell means
}
\description{
Simulate esco trajetory cell means
}
\details{
Cells are first assigned to a path and a step along that path. This is
controlled by the \code{cells.design} parameter which is a \code{data.frame}
with the columns "traj", "Probability", "Alpha" and "Beta". The Path field
is an ID for each path and the Probability field is the probability that a
cell will come from that path (must sum to 1). The Alpha and Beta parameters
control the density of cells along the path. After they are assigned to paths
the step for each cell is sampled from a Beta distribution with parameters
shape1 equals Alpha and shape2 equals beta. The distribution
can be viewed using \code{hist(rbeta(10000, Alpha, Beta), breaks = 100)}.
Once cells are assigned to paths and steps the correct means are extracted
from the \code{paths.means} parameter and adjusted based on each cell's
library size. An adjustment for BCV is then applied.
}
