% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimTreeCellMeans}
\alias{escoSimTreeCellMeans}
\title{Simulate tree cell means}
\usage{
escoSimTreeCellMeans(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment containing simulation.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
sim SingleCellExperiment with cell means
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell where the cells are of tree structure. For a cell we start with the base gene mean, 
and then combined with the DE factor if the gene is DE gene. 
Additional factors are applied if the DE gene are also marker genes. The selected
means are then adjusted for each cell's expected library size.
}
