% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimTreeDE}
\alias{escoSimTreeDE}
\title{Simulate differential expression (DE) factors for tree structure}
\usage{
escoSimTreeDE(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add differential expression to.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SingleCellExperiment with simulated differential expression.
}
\description{
DE factors for each tree branches are produced using \code{\link{getLNormFactors}} with 
means follows a multivariate normal ditribution, where the correlation relationship is generated from the tree structure.
}
\details{
The method of simulating paths is inspired by the method used in SymSim
simulation. Particularly, the tree structure is generated by making the 
DE factors more similar for closer tree branches,
and less similar for far away tree branches. The tree structure is input through \code{tree} in object \code{escoParams}, 
and we generate the relationship between DE factors for different branches from the estimated evolutionary 
variance-covariance matrix of the tree using \code{\link{vcv.phylo}}.
}
\references{
Zhang, X., Xu, C. & Yosef, N. Simulating multiple faceted variability in single cell RNA sequencing. 
Nat Commun 10, 2611 (2019). \url{https://doi.org/10.1038/s41467-019-10500-w}.
}
