% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimTrueCounts}
\alias{escoSimTrueCounts}
\title{Simulate true counts}
\usage{
escoSimTrueCounts(sim, type, verbose, numCores = 2)
}
\arguments{
\item{sim}{SingleCellExperiment to add true counts to.}

\item{type}{The type of cell structure: 'single', 'group', 'tree', or 'traj'.}

\item{verbose}{whether to print the process or not}

\item{numCores}{the number of cores for parallelization during simulation, if set as NULL, then all the cores are used.}
}
\value{
sim SingleCellExperiment with simulated true counts.
}
\description{
Simulate a true counts matrix. Counts are simulated from a poisson
distribution where each gene in each cell has it's own mean based on the
group (or path position), expected library size and BCV. Additional gene-gene 
interaction can be introduced via a copula model with correlation matrix input 
by design or randomly generated from a real purifed gene expression data set from NeuroExpresso
\url{https://www.eurekalert.org/pub_releases/2017-11/sfn-nwa111417.php}.
}
