% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimTrueCounts}
\alias{escoSimTrueCounts}
\title{Simulate true counts}
\usage{
escoSimTrueCounts(sim, type, verbose, numCores = 2)
}
\arguments{
\item{sim}{SingleCellExperiment to add true counts to.}
}
\value{
SingleCellExperiment with simulated true counts.
}
\description{
Simulate a true counts matrix. Counts are simulated from a poisson
distribution where Each gene in each cell has it's own mean based on the
group (or path position), expected library size and BCV.
}
