% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimZeroInflate}
\alias{escoSimZeroInflate}
\title{Simulate noisy observed count matirx via adding more zeros through a zero inflation model}
\usage{
escoSimZeroInflate(sim, trial, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add dropout to.}

\item{trial}{the index of trial of simulation, all trials share the same truth but the 
noise is added independently for each trial}

\item{verbose}{whether to print the process or not}
}
\value{
SingleCellExperiment with simulated dropout and observed counts.
}
\description{
A logistic function is used to form a relationshop between the expression
level of a gene and the probability of dropout, giving a probability for each
gene in each cell. These probabilities are used in a Bernoulli distribution
to decide which counts should be dropped.
}
