% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{expandbinary}
\alias{expandbinary}
\title{expand transcript counts to a vector of binaries of the same length of as the number of transcripts}
\usage{
expandbinary(true_counts_1cell)
}
\arguments{
\item{true_counts_1cell}{number of transcript in one cell}
}
\description{
expand transcript counts to a vector of binaries of the same length of as the number of transcripts
}
