% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escoParams-methods.R
\docType{methods}
\name{expandParams,escoParams-method}
\alias{expandParams,escoParams-method}
\title{this function is borrowed from splatter}
\usage{
\S4method{expandParams}{escoParams}(object)
}
\description{
this function is borrowed from splatter
}
