% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{expandParams,Params-method}
\alias{expandParams,Params-method}
\title{this function is borrowed from Splatter}
\usage{
\S4method{expandParams}{Params}(object, vectors, n)
}
\description{
this function is borrowed from Splatter
}
