% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params-methods.R, R/escoParams-methods.R
\docType{methods}
\name{expandParams}
\alias{expandParams}
\alias{expandParams,Params-method}
\alias{expandParams,escoParams-method}
\title{Expand values of parameters}
\usage{
expandParams(object, vectors, n)

\S4method{expandParams}{Params}(object, vectors, n)

\S4method{expandParams}{escoParams}(object)
}
\arguments{
\item{object}{Params object to set parameters in.}

\item{vectors}{list of parameters to reset values,
where \code{names(update)} are the
      names of the parameters to set and the items 
      in the list are values.}

\item{n}{an integer indicates the target number of repititions.}
}
\value{
object with new parameters.
}
\description{
Expand the values of multiple parameters by repititions in a Params object.
This function is similar with that in \code{\link[splatter]{splatter}}.
}
\details{
Each parameter is set by a call to \code{\link{setParam}}. 
If the same parameter is specified multiple times 
it will be set multiple times.
}
