% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{getBetaStepProbs}
\alias{getBetaStepProbs}
\title{Get Beta step probabilities}
\usage{
getBetaStepProbs(steps, alpha, beta)
}
\arguments{
\item{steps}{Number of steps}

\item{alpha}{Alpha parameter}

\item{beta}{Beta parameter}
}
\value{
Vector of probabilities
}
\description{
Use a Beta distribution for set probabilities along a path (this function is borrowed from splatter)
}
\details{
The density is sampled from a Beta distribution between 0 and 1. Infinite
densities at edges are adjusted and then the values are scaled to give
probabilities.
}
