% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{getLNormFactors}
\alias{getLNormFactors}
\title{Get log-normal factors}
\usage{
getLNormFactors(n.facs, is.selected, neg.prob = 0, fac.loc, fac.scale)
}
\arguments{
\item{n.facs}{Number of factors to generate.}

\item{neg.prob}{Probability that a selected factor is less than one.}

\item{fac.loc}{Location parameter for the log-normal distribution.}

\item{fac.scale}{Scale factor for the log-normal distribution.}

\item{sel.prob}{Probability that a factor will be selected to be different
from 1.}
}
\value{
Vector containing generated factors.
}
\description{
Randomly generate multiplication factors from a log-normal distribution.
}
