% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getParam,Params-method}
\alias{getParam,Params-method}
\title{this function is borrowed from Splatter}
\usage{
\S4method{getParam}{Params}(object, name)
}
\description{
this function is borrowed from Splatter
}
