% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params-methods.R
\docType{methods}
\name{getParams}
\alias{getParams}
\alias{getParams,Params-method}
\title{Get parameters}
\usage{
getParams(object, names)

\S4method{getParams}{Params}(object, names)
}
\arguments{
\item{object}{Params object to get values from.}

\item{names}{vector of names of the parameters to get.}
}
\value{
List with the values of the selected parameters.
}
\description{
Get multiple parameter values from a Params object. 
This function is similar with that in \code{\link[splatter]{splatter}}.
}
\examples{
params <- newescoParams()
getParams(params, c("nGenes", "nCells"))
}
