% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R, R/esco-simulate.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(x, x0, k)

logistic(x, x0, k)
}
\arguments{
\item{x}{value to apply the function to.}

\item{x0}{midpoint parameter. Gives the centre of the function.}

\item{k}{shape parameter. Gives the slope of the function.}

\item{x}{value to apply the function to.}

\item{x0}{midpoint parameter. Gives the centre of the function.}

\item{k}{shape parameter. Gives the slope of the function.}
}
\value{
Value of logistic funciton with given parameters

Value of logistic funciton with given parameters
}
\description{
Implementation of the logistic function

Implementation of the logistic function
}
