% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makespd}
\alias{makespd}
\title{Make positive definite 
Make a matrix positive definite  through tweaking the eigen values}
\usage{
makespd(rho)
}
\arguments{
\item{rho}{the input matrix (assuming squared matrix)}
}
\value{
a positive definite matrix
}
\description{
Make positive definite 
Make a matrix positive definite  through tweaking the eigen values
}
