% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/escoParams-methods.R
\name{newParams}
\alias{newParams}
\alias{newescoParams}
\title{New Params}
\usage{
newescoParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
New Params object.
}
\description{
Create a new Params object. Functions exist for each of the different
Params subtypes.
}
\examples{
params <- newSimpleParams()
params <- newSimpleParams(nGenes = 200, nCells = 10)

}
