% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{Observed_Counts}
\alias{Observed_Counts}
\title{Simulate observed count matrix given technical biases and the true counts}
\usage{
Observed_Counts(sim, trial, verbose, protocol = "UMI", nbatch = 1)
}
\arguments{
\item{protocol}{a string, can be "nonUMI" or "UMI"}

\item{nbatch}{number of batches}

\item{true_counts}{gene cell matrix}

\item{meta_cell}{the meta information related to cells, will be combined with technical cell level information and returned}

\item{alpha_mean}{the mean of rate of subsampling of transcripts during capture step, default at 10 percent efficiency}

\item{alpha_sd}{the std of rate of subsampling of transcripts}

\item{lenslope}{amount of length bias}

\item{nbins}{number of bins for gene length}

\item{amp_bias_limit}{range of amplification bias for each gene, a vector of length ngenes}

\item{rate_2PCR}{PCR efficiency, usually very high, default is 0.8}

\item{nPCR1}{the number of PCR cycles, default is 16}

\item{LinearAmp}{if linear amplification is used for pre-amplification step, default is FALSE}

\item{LinearAmp_coef}{the coeficient of linear amplification, that is, how many times each molecule is amplified by}

\item{depth_mean}{mean of sequencing depth}

\item{depth_sd}{std of sequencing depth}

\item{hge2true}{if we add high gene expression to true counts}

\item{SE}{input, should be a summerized experiment rather than a list of elements, default is False}
}
\description{
Simulate observed count matrix given technical biases and the true counts
}
