% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puri_data}
\alias{puri_data}
\title{a purified gene expression data set to sample correlation strcuture from}
\format{A matrix of 2000 rows (genes) and 256 columns (cells), where each entry is 
the log transformed normalized gene expression.}
\source{
\url{https://www.eurekalert.org/pub_releases/2017-11/sfn-nwa111417.php}
}
\usage{
data(puri_data)
}
\description{
a purified gene expression data set to sample correlation strcuture from
}
\examples{
data(puri_data)
}
\keyword{datasets}
