% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randcop}
\alias{randcop}
\title{Sample from a gaussian copula}
\usage{
randcop(Rho, nCells)
}
\arguments{
\item{Rho}{the correlation matrix in the gaussian copula}

\item{nCells}{the number of samples to draw}
}
\value{
matrix of nCell rows, where each row is a sample from the gaussian copula
}
\description{
Sample from a gaussian copula
}
