% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randcor}
\alias{randcor}
\title{Random sampeled correlation from the reallife gene correlations}
\usage{
randcor(ngenes)
}
\arguments{
\item{ngenes}{the number of genes}
}
\value{
a correlation matrix of dimension ngenes*ngenes
}
\description{
Generate a random correlation matrix that is positive definite, using a purified gene expression data set 
\url{https://www.eurekalert.org/pub_releases/2017-11/sfn-nwa111417.php}.
}
