% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RealDropParams}
\alias{RealDropParams}
\alias{RealDropParams-class}
\title{The RealDropParams class}
\description{
S4 class that holds parameters for the RealDrop simulation.
}
\section{Parameters}{


The RealDrop simulation requires the following parameters:

\describe{
    \item{\code{nGenes}}{The number of genes to simulate.}
    \item{\code{nCells}}{The number of cells to simulate.}
    \item{\code{[seed]}}{Seed to use for generating random numbers.}
    \item{\emph{Dropout parameters}}{
        \describe{
            \item{\code{dropout.mid}}{Midpoint parameter for the dropout
            logistic function.}
            \item{\code{dropout.shape}}{Shape parameter for the dropout
            logistic function.}
        }
    }
}

For details of the RealDrop simulation
see \code{\link{RealDropSimulate}}.
}

