% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{rnorm_truc}
\alias{rnorm_truc}
\title{sample from truncated normal distribution}
\usage{
rnorm_truc(n, mean, sd, a, b)
}
\arguments{
\item{a}{the minimum value allowed}

\item{b}{the maximum value allowed}
}
\description{
sample from truncated normal distribution
}
