% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escoParams-methods.R
\docType{methods}
\name{setParams,escoParams-method}
\alias{setParams,escoParams-method}
\title{this function is borrowed from splatter}
\usage{
\S4method{setParams}{escoParams}(object, update = NULL, ...)
}
\description{
this function is borrowed from splatter
}
