% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setParams,Params-method}
\alias{setParams,Params-method}
\title{this function is borrowed from Splatter}
\usage{
\S4method{setParams}{Params}(object, update = NULL, ...)
}
\description{
this function is borrowed from Splatter
}
