% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setParams}
\alias{setParams}
\title{this function is borrowed from Splatter
Set parameters}
\usage{
setParams(object, update = NULL, ...)
}
\description{
this function is borrowed from Splatter
Set parameters
}
