% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params-methods.R
\docType{methods}
\name{setParamsUnchecked}
\alias{setParamsUnchecked}
\alias{setParamsUnchecked,Params-method}
\title{Set parameters UNCHECKED}
\usage{
setParamsUnchecked(object, update = NULL, ...)

\S4method{setParamsUnchecked}{Params}(object, update = NULL, ...)
}
\arguments{
\item{object}{Params object to set parameters in.}

\item{update}{list of parameters to set where \code{names(update)} are the
names of the parameters to set and the items in the list are values.}

\item{...}{additional parameters to set. These are combined with any
parameters specified in \code{update}.}
}
\value{
Params object with updated values.
}
\description{
Set multiple parameters in a Params object without checking for validity.
This function is similar with that in \code{\link[splatter]{splatter}}.
}
\details{
Each parameter is set by a call to \code{\link{setParam}}. If the same
parameter is specified multiple times it will be set multiple times.
Parameters can be specified using a list via \code{update} (useful when
collecting parameter values in some way) or individually (useful when setting
them manually), see examples.
}
