% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setParamUnchecked,Params-method}
\alias{setParamUnchecked,Params-method}
\title{this function is borrowed from Splatter}
\usage{
\S4method{setParamUnchecked}{Params}(object, name, value)
}
\description{
this function is borrowed from Splatter
}
