% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setParamUnchecked}
\alias{setParamUnchecked}
\title{this function is borrowed from Splatter
Set a parameter UNCHECKED}
\usage{
setParamUnchecked(object, name, value)

setParamUnchecked(object, name, value)
}
\description{
this function is borrowed from Splatter
Set a parameter UNCHECKED

this function is borrowed from Splatter
Set a parameter UNCHECKED
}
