% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,Params-method}
\alias{show,Params-method}
\title{this function is borrowed from Splatter}
\usage{
\S4method{show}{Params}(object)
}
\description{
this function is borrowed from Splatter
}
