% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{splatEstCorr}
\alias{splatEstCorr}
\title{Estimate Splat correlation matrix using meta data or the copula method}
\usage{
splatEstCorr(counts, cellinfo, params, method = c("meta", "copula"))
}
\description{
Estimate Splat correlation matrix using meta data or the copula method
}
