% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-estimate.R
\name{escoEstimate}
\alias{escoEstimate}
\alias{escoEstimate.SingleCellExperiment}
\alias{escoEstimate.matrix}
\title{Estimate esco simulation parameters}
\usage{
escoEstimate(counts, dirname, group = FALSE, cellinfo = NULL,
  params = newescoParams())

\method{escoEstimate}{SingleCellExperiment}(counts, dirname,
  group = FALSE, cellinfo = NULL, params = newescoParams())

\method{escoEstimate}{matrix}(counts, dirname, group = FALSE,
  cellinfo = NULL, params = newescoParams())
}
\arguments{
\item{counts}{either a counts matrix or a SingleCellExperiment object
containing count data to estimate parameters from.}

\item{dirname}{a tring of directory name to indicate where to save the results}

\item{group}{whether the data is believed to be of discrete cell groups or not, 
if yes, the corresponding cellinfo indicating the cell group labels need to be input as well}

\item{cellinfo}{a vector of length n, where n is the number of cells. 
Each entries is the group identity of a cell.}

\item{params}{escoParams object to store estimated values in.}
}
\value{
escoParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the esco simulation from a real
dataset. See the individual estimation functions for more details on how this
is done.
}
\examples{
# Load example data
library(scater)
data("sc_example_counts")

params <- escoEstimate(sc_example_counts)
params
}
\seealso{
\code{\link{escoEstMean}},  \code{\link{escoEstLib}},
\code{\link{escoEstOutlier}}, \code{\link{escoEstBCV}},
\code{\link{escoEstDropout}}, \code{\link{escoEstDE}},
\code{\link{escoEstGroupMean}}, \code{\link{escoEstGroupLib}},
\code{\link{escoEstGroupOutlier}}
}
