% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn.R
\name{gcn_error}
\alias{gcn_error}
\title{Compute the estiamtion error of gene co-expression networks}
\usage{
gcn_error(gcnlist, method = c("MSE", "Spectral", "Inf", "L1", "ARI",
  "Jaccard"), clustk = 9, clustmethod = "hclust")
}
\arguments{
\item{gcnlist}{a list of gene coexpression networks of the same set of genes, 
where the first is taken as the truth, and the rest are all estimations.}

\item{method}{a vector of error measure names, possible choices are mean square error "MSE", 
Spectral norm "Spectral", infinite norm "Inf", l1 norm, "L1", Adjusted rand index \code{\link{adjustedRand}} with 
\code{clusk} numbers of clusters "ARI", Jaccard index with 
\code{clusk} numbers of clusters "Jaccard".}

\item{clustk}{the number of clusters used in "ARI" and "Jaccard" error measure. Default is 9.}

\item{clustmethod}{what clustering method to use. Possible choices are hirarchical clustering "hclust" \code{\link{hclust}}, 
and "kmeans" \code{\link{kmeans}}. Default is "hclust".}
}
\value{
a matrix of error levels, where each row indicates the given different estimation of gcn, 
        and each column indicates different error measure.
}
\description{
Compute the error between the true and estimated gene co-expression network using different error measure
}
\examples{
data1 = matrix(rnorm(100),20,5)
data2 = matrix(rnorm(100),20,5)
gcnlist = list(cor(data1),cor(data2))
gcnerr = gcn_error(gcnlist)
}
