% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimSingleCellMeans}
\alias{escoSimSingleCellMeans}
\alias{escoSimGroupCellMeans}
\alias{escoSimTreeCellMeans}
\alias{escoSimPathCellMeans}
\title{Simulate cell means}
\usage{
escoSimSingleCellMeans(sim, verbose)

escoSimGroupCellMeans(sim, verbose)

escoSimTreeCellMeans(sim, verbose)

escoSimPathCellMeans(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add cell means to.}

\item{verbose}{logical. Whether to print progress messages}

\item{sim}{SingleCellExperiment containing simulation.}

\item{params}{escoParams object with simulation parameters.}
}
\value{
SingleCellExperiment with added cell means.

SingleCellExperiment with cell means
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell. Cells start with the mean expression for the group they belong to
(when simulating groups) or cells are assigned the mean expression from a
random position on the appropriate path (when simulating tree). The selected
means are adjusted for each cell's expected library size.
}
\details{
Cells are first assigned to a path and a step along that path. This is
controlled by the \code{cells.design} parameter which is a \code{data.frame}
with the columns "traj", "Probability", "Alpha" and "Beta". The Path field
is an ID for each path and the Probability field is the probability that a
cell will come from that path (must sum to 1). The Alpha and Beta parameters
control the density of cells along the path. After they are assigned to paths
the step for each cell is sampled from a Beta distribution with parameters
shape1 equals Alpha and shape2 equals beta. The distribution
can be viewed using \code{hist(rbeta(10000, Alpha, Beta), breaks = 100)}.
Once cells are assigned to paths and steps the correct means are extracted
from the \code{paths.means} parameter and adjusted based on each cell's
library size. An adjustment for BCV is then applied.
}
