% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimDE}
\alias{escoSimDE}
\alias{escoSimGroupDE}
\alias{escoSimTreeDE}
\title{Simulate group differential expression}
\usage{
escoSimGroupDE(sim, verbose)

escoSimTreeDE(sim, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment to add differential expression to.}
}
\value{
SingleCellExperiment with simulated differential expression.
}
\description{
Simulate differential expression. Differential expression factors for each
group are produced using \code{\link{getLNormFactors}} and these are added
along with updated means for each group. For paths care is taked to make sure
they are simulated in the correct order.
}
