% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco-simulate.R
\name{escoSimDownSample}
\alias{escoSimDownSample}
\title{Simulate observed count matrix though downsampling.
(This function is borrowed from SymSim).}
\usage{
escoSimDownSample(sim, trial, verbose, numCores = 2, nbatch = 1)
}
\arguments{
\item{sim}{SingleCellExperiment to add dropout to.}

\item{trial}{the index of trial of simulation, all trials share the same truth but the 
noise is added independently for each trial}

\item{verbose}{whether to print the process or not}

\item{numCores}{the number of cores used for parallelization, default is 2.
If set as NULL, then all the detected cores are used.}

\item{nbatch}{number of batches, default is 1.}
}
\value{
SingleCellExperiment with simulated dropout and observed counts.
}
\description{
Simulate observed count matrix though downsampling.
(This function is borrowed from SymSim).
}
\references{
Zhang X, Xu C, Yosef N. Simulating multiple faceted variability in single cell RNA sequencing. 
Nature communications. 2019 Jun 13;10(1):1-6. 
\url{https://www.nature.com/articles/s41467-019-10500-w}
}
