% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{expandParams}
\alias{expandParams}
\title{this function is borrowed from Splatter
Expand parameters}
\usage{
expandParams(object, ...)
}
\description{
this function is borrowed from Splatter
Expand parameters
}
