% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/escoParams-methods.R
\docType{methods}
\name{expandParams}
\alias{expandParams}
\alias{expandParams,Params-method}
\alias{expandParams,escoParams-method}
\title{Expand parameters}
\usage{
expandParams(object, ...)

\S4method{expandParams}{Params}(object, vectors, n)

\S4method{expandParams}{escoParams}(object)
}
\value{
object with new parameters.
}
\description{
Expand parameters

Expand multiple entries esco parameter object
}
