% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params-methods.R
\docType{methods}
\name{getParam}
\alias{getParam}
\alias{getParam,Params-method}
\title{Get a parameter}
\usage{
getParam(object, name)

\S4method{getParam}{Params}(object, name)
}
\arguments{
\item{object}{Params object to get values from.}

\item{name}{vector of names of the parameters to get.}
}
\value{
List with the values of the selected parameters.
}
\description{
Get a parameter value from a Params object.  
This function is similar with that in \code{\link[splatter]{splatter}}.
}
\examples{
params <- newescoParams()
getParam(params, c("nGenes"))
}
