% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/escoParams-methods.R
\name{newParams}
\alias{newParams}
\alias{newescoParams}
\title{New Params}
\usage{
newescoParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}

\item{...}{any additional parameter settings to override what is provided in
\code{\link{params}}.}
}
\value{
New Params object.

Object with new parameter value.
}
\description{
Create a new Params object. Functions exist for each of the different
Params subtypes.

To initiate a new object for storing esco simulation parameters and process.
The default parameters are set as in \code{escoParams}.
}
\examples{
params <- newescoParams()
params <- newescoParams(nGenes = 200, nCells = 10)

para = newescoParams()
}
