% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escoParams-methods.R
\name{newescoParams}
\alias{newescoParams}
\title{Initiate a new escoParam object}
\usage{
newescoParams(...)
}
\arguments{
\item{...}{any additional parameter settings to override what is provided in
\code{\link{params}}.}
}
\value{
Object with new parameter value.
}
\description{
To initiate a new object for storing esco simulation parameters and process.
The default parameters are set as in \code{escoParams}.
}
\examples{
para = newescoParams()
}
