% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params-methods.R, R/escoParams-methods.R
\docType{methods}
\name{setParam}
\alias{setParam}
\alias{setParam,Params-method}
\alias{setParam,escoParams-method}
\title{Set a parameter}
\usage{
setParam(object, name, value)

\S4method{setParam}{Params}(object, name, value)

\S4method{setParam}{escoParams}(object, name, value)
}
\arguments{
\item{object}{object to set parameter in.}

\item{name}{name of the parameter to set.}

\item{value}{value to set the parameter to.}
}
\value{
Object with new parameter value.
}
\description{
Function for setting a parameter value. 
This function is similar with that in \code{\link[splatter]{splatter}}.
}
\examples{
params <- newescoParams()
setParam(params, "nGenes", 100)

}
