% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setParamUnchecked}
\alias{setParamUnchecked}
\title{Set a parameter UNCHECKED}
\usage{
setParamUnchecked(object, name, value)

setParamUnchecked(object, name, value)
}
\description{
Set a parameter UNCHECKED

Set a parameter UNCHECKED
}
